+++ web/src/constants/signUpConstants.ts
@@ -1,0 +1,7 @@
+export const SIGNUP_HEADER_CONFIG: Record<number, string> = {
+  1: "닉네임 입력",
+  2: "기본 정보 선택",
+  3: "기본 정보 선택",
+  4: "거래 방식",
+  5: "주요 거래 품목",
+};
